회원 검색 기능 추가 전

+++ src/main/java/com/multi/webiting/AdminController.java
@@ -3,6 +3,8 @@ package com.multi.webiting;
 import java.util.List;
 
 import javax.inject.Inject;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
 
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Controller;
@@ -10,9 +12,11 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
+import com.board.model.PagingVO;
 import com.common.CommonUtil;
 import com.user.model.UserVO;
 import com.user.service.AdminService;
@@ -29,45 +33,77 @@ public class AdminController {
 	
 	CommonUtil common = new CommonUtil();
 	
-	@GetMapping("/userList")
-	public String userList(Model m) { //, @ModelAttribute("page") PagingVO page, HttpServletRequest req, @RequestHeader("User-Agent") String userAgent
-		/*
-		 * HttpSession ses=req.getSession(); log.info("1. page===="+page); int
-		 * totalCount=this.aService.getTotalCount(); page.setTotalCount(totalCount);
-		 * page.setPagingBlock(10); page.init(ses); log.info("2. page===="+page);
-		 * List<UserVO> userArr=aService.listUser(null); String loc="admin/list"; String
-		 * pageNavi=page.getPageNavi("/", loc, userAgent);
-		 * 
-		 * m.addAttribute("pageNavi", pageNavi); m.addAttribute("paging", page);
-		 * m.addAttribute("userArr", userArr); return "member/userlist";
-		 */
-		
+	@GetMapping("/userList_old")
+	public String userList(Model m) {
 		int totalCount=aService.getTotalCount();
-		List<UserVO> arr=aService.listUser();
+		//List<UserVO> arr=aService.listUser();
 		
-		m.addAttribute("userArr",arr);		
+		//m.addAttribute("userArr",arr);		
 		m.addAttribute("totalCount",totalCount);
 		
 		return "member/userlist";
 	}
 	
+	//PagingVO에 getPageNavi()추가 (paging java로 처리)ㅡ 검색 기능 추가
+	@GetMapping("/userList")
+	public String userListPaging(Model m, @ModelAttribute("page") PagingVO page,
+			HttpServletRequest req, @RequestHeader("User-Agent") String userAgent) {
+		String myctx=req.getContextPath();
+		
+		HttpSession ses=req.getSession();
+		
+		log.info("1. page===="+page);
+		int totalCount=this.aService.getTotalCount(page);
+		page.setTotalCount(totalCount);
+		page.setPagingBlock(5);
+		page.init(ses);
+		
+		log.info("2. page===="+page);
+		List<UserVO> userArr=this.aService.selectUserAllPaging(page);
+		String loc="admin/userList";
+		String pageNavi=page.getPageNavi(myctx, loc, userAgent);
+		
+		m.addAttribute("pageNavi", pageNavi);
+		m.addAttribute("paging", page);
+		m.addAttribute("userArr", userArr);
+		return "member/userlist";
+		
+	}
+	
 	
 	@PostMapping("/userDel")
-	public String userDelete(@RequestParam(defaultValue="0") int idx) {
+	public String userDelete(Model m, @RequestParam(defaultValue="0") int idx) {
 		log.info("idx==="+idx);
+		
 		if(idx==0) {
 			return "redirect:userList";
 		}
+		
+		UserVO vo=this.aService.selectUserByIdx(idx);
+		if(vo==null) {
+			return common.addMsgBack(m, "해당 회원은 존재하지 않습니다.");
+		}
+		
 		int n=aService.deleteUser(idx);
 		
-		return "redirect:userList";
+		String str=(n>0)?"삭제 되었습니다.":"회원 삭제 실패";
+		String loc=(n>0)?"userList":"javascript:history.back()";
+		
+		return common.addMsgLoc(m, str, loc);
 	}
 	
 	@PostMapping("/userEdit")
-	public String userEdit(Model m, @RequestParam int idx) {
+	public String userEdit(Model m, @RequestParam(defaultValue="0") int idx) {
 		log.info("idx==="+idx);
 		
+		if(idx==0) {
+			return "redirect:userList";
+		}
+		
 		UserVO vo=this.aService.selectUserByIdx(idx);
+		if(vo==null) {
+			return common.addMsgBack(m, "해당 회원은 존재하지 않습니다.");
+		}
 		
 		m.addAttribute("user", vo);
